expert_agent/api/requirements.txt
python-dotenv
redis
chromadb
langchain
langchain-community
langchain-text-splitters
langchain-experimental
langchain-openai>=0.1.0
openpyxl
python-pptx
pillow
openai
pandas
